Criação da Library de pagamento
